refactor: Split tui/ui.rs into tool_renderer.rs and message_formatter.rs
Extract tool/JSON rendering and message formatting into focused modules.
Reduces ui.rs from 1703 LOC to 1078 LOC (37% reduction).

- message_formatter.rs (361 LOC): User/assistant/system message formatting,
  text wrapping utilities, line height calculation
- tool_renderer.rs (306 LOC): Tool call display, JSON tree rendering,
  parameter/response formatting

Fixes #394
