+++ crates/cli/src/tui/ui.rs
@@ -1,4 +1,8 @@
 //! Rendering layer for TUI - pure functions, no state mutation
+//!
+//! This module coordinates rendering of the TUI by delegating to:
+//! - `message_formatter`: User/assistant/system message formatting
+//! - `tool_renderer`: Tool call and JSON parameter rendering
 
 use ratatui::{
     layout::Rect,
@@ -15,10 +19,13 @@ use unicode_width::UnicodeWidthStr;
 use crate::commands::permissions_ui;
 use crate::tui::app::{App, LayoutCache};
 use crate::tui::layout::{LayoutConfig, LayoutOrganizer};
-use crate::tui::message::{Message, Role};
+use crate::tui::message::Role;
+
+use super::message_formatter::{calculate_wrapped_height, format_message_by_role};
+use super::tool_renderer::{format_response_content, format_tool_parameters, format_tool_params};
 
 /// Rust-themed colors for structural elements (borders, titles)
-const RUST_ORANGE: Color = Color::Rgb(222, 165, 132);
+pub(super) const RUST_ORANGE: Color = Color::Rgb(222, 165, 132);
 
 /// Main render function - updates TextArea block style for focus-aware rendering
 /// Returns (max_scroll, debug_max_scroll, layout_cache) tuple for app state update
@@ -249,634 +256,6 @@ fn render_status_bar(frame: &mut Frame, area: Rect, app: &mut App, throbber: cha
     let status = Paragraph::new(Line::from(status_spans)).style(Style::default().bg(Color::Black));
 
     frame.render_widget(status, area);
-}
-
-/// Format a tool call for display (compact JSON parameters)
-/// Max width parameter controls truncation to fit header
-fn format_tool_params(params: &serde_json::Value, max_width: usize) -> String {
-    let formatted = match params {
-        serde_json::Value::Object(map) if map.is_empty() => "()".to_string(),
-        serde_json::Value::Object(map) => {
-            let items: Vec<String> = map
-                .iter()
-                .take(3) // Show max 3 params
-                .map(|(k, v)| {
-                    let value_str = match v {
-                        serde_json::Value::String(s) if s.len() > 30 => {
-                            format!("\"{}...\"", &s[..27])
-                        }
-                        serde_json::Value::String(s) => format!("\"{}\"", s),
-                        _ => v.to_string(),
-                    };
-                    format!("{}: {}", k, value_str)
-                })
-                .collect();
-
-            if map.len() > 3 {
-                format!("({}, ...)", items.join(", "))
-            } else {
-                format!("({})", items.join(", "))
-            }
-        }
-        _ => format!("({})", params),
-    };
-
-    // Truncate to max_width if needed
-    if formatted.len() > max_width {
-        format!("{}...)", &formatted[..max_width.saturating_sub(4)])
-    } else {
-        formatted
-    }
-}
-
-/// Recursively format a JSON value with tree structure
-/// Returns formatted lines with proper indentation
-fn format_json_value(
-    value: &serde_json::Value,
-    parent_bar: &str,
-    tree_char: &str,
-    continuation_char: &str,
-    key: Option<&str>,
-    content_width: usize,
-    depth: usize,
-) -> Vec<Line<'static>> {
-    let mut lines = Vec::new();
-
-    // Safety limit for recursion depth
-    if depth > 10 {
-        let label = key.map(|k| format!("{}: ", k)).unwrap_or_default();
-        lines.push(Line::from(Span::styled(
-            format!("{} {} {}[max depth exceeded]", parent_bar, tree_char, label),
-            Style::default().fg(Color::DarkGray),
-        )));
-        return lines;
-    }
-
-    match value {
-        serde_json::Value::Object(map) => {
-            // Object - show as nested structure
-            let label = key.map(|k| format!("{}: ", k)).unwrap_or_default();
-            lines.push(Line::from(Span::styled(
-                format!("{} {} {}", parent_bar, tree_char, label),
-                Style::default().fg(Color::DarkGray),
-            )));
-
-            // Nested items
-            let nested_parent_bar = format!("{} {} ", parent_bar, continuation_char);
-            for (idx, (nested_key, nested_value)) in map.iter().enumerate() {
-                let is_last = idx == map.len() - 1;
-                let nested_tree_char = if is_last { "└─" } else { "├─" };
-                let nested_continuation_char = if is_last { "  " } else { "│ " };
-
-                let nested_lines = format_json_value(
-                    nested_value,
-                    nested_parent_bar.trim_end(),
-                    nested_tree_char,
-                    nested_continuation_char,
-                    Some(nested_key),
-                    content_width,
-                    depth + 1,
-                );
-                lines.extend(nested_lines);
-            }
-        }
-        serde_json::Value::Array(arr) => {
-            // Array - show as indexed items
-            let label = key.map(|k| format!("{}: ", k)).unwrap_or_default();
-            lines.push(Line::from(Span::styled(
-                format!(
-                    "{} {} {}[{} items]",
-                    parent_bar,
-                    tree_char,
-                    label,
-                    arr.len()
-                ),
-                Style::default().fg(Color::DarkGray),
-            )));
-
-            // Array items
-            let nested_parent_bar = format!("{} {} ", parent_bar, continuation_char);
-            for (idx, item) in arr.iter().enumerate() {
-                let is_last = idx == arr.len() - 1;
-                let nested_tree_char = if is_last { "└─" } else { "├─" };
-                let nested_continuation_char = if is_last { "  " } else { "│ " };
-                let item_key = format!("[{}]", idx);
-
-                let nested_lines = format_json_value(
-                    item,
-                    nested_parent_bar.trim_end(),
-                    nested_tree_char,
-                    nested_continuation_char,
-                    Some(&item_key),
-                    content_width,
-                    depth + 1,
-                );
-                lines.extend(nested_lines);
-            }
-        }
-        _ => {
-            // Primitive value - format with wrapping
-            let value_str = match value {
-                serde_json::Value::String(s) => s.clone(),
-                serde_json::Value::Number(n) => n.to_string(),
-                serde_json::Value::Bool(b) => b.to_string(),
-                serde_json::Value::Null => "null".to_string(),
-                _ => unreachable!(),
-            };
-
-            let label = key.map(|k| format!("{}: ", k)).unwrap_or_default();
-            let first_line_prefix = format!("{} {} {}", parent_bar, tree_char, label);
-            let continuation_prefix = format!("{} {}  ", parent_bar, continuation_char);
-            let wrapped_lines = wrap_value_with_indent(
-                &value_str,
-                &first_line_prefix,
-                &continuation_prefix,
-                content_width,
-            );
-
-            for wrapped_line in wrapped_lines {
-                lines.push(Line::from(Span::styled(
-                    wrapped_line,
-                    Style::default().fg(Color::DarkGray),
-                )));
-            }
-        }
-    }
-
-    lines
-}
-
-/// Wrap text with proper indentation for continuation lines
-/// Used for long parameter/response values that need to wrap
-fn wrap_value_with_indent(
-    text: &str,
-    first_line_prefix: &str,
-    continuation_prefix: &str,
-    max_width: usize,
-) -> Vec<String> {
-    if text.is_empty() {
-        return vec![format!("{}", first_line_prefix)];
-    }
-
-    // Calculate available width for first line
-    let first_width = max_width.saturating_sub(first_line_prefix.len()).max(20);
-    // Continuation lines get full continuation_prefix width
-    let cont_width = max_width.saturating_sub(continuation_prefix.len()).max(20);
-
-    let mut result = Vec::new();
-    let words: Vec<&str> = text.split_whitespace().collect();
-
-    if words.is_empty() {
-        result.push(first_line_prefix.to_string());
-        return result;
-    }
-
-    let mut current_line = String::new();
-    let mut is_first = true;
-
-    for word in words {
-        let test_line = if current_line.is_empty() {
-            word.to_string()
-        } else {
-            format!("{} {}", current_line, word)
-        };
-
-        let width_limit = if is_first { first_width } else { cont_width };
-
-        if test_line.len() <= width_limit {
-            current_line = test_line;
-        } else {
-            // Current line is full, push it
-            if !current_line.is_empty() {
-                let prefix = if is_first {
-                    first_line_prefix
-                } else {
-                    continuation_prefix
-                };
-                result.push(format!("{}{}", prefix, current_line));
-                is_first = false;
-            }
-            current_line = word.to_string();
-        }
-    }
-
-    // Push final line
-    if !current_line.is_empty() {
-        let prefix = if is_first {
-            first_line_prefix
-        } else {
-            continuation_prefix
-        };
-        result.push(format!("{}{}", prefix, current_line));
-    }
-
-    result
-}
-
-/// Format tool parameters in a clean, readable way (not raw JSON)
-/// Returns tree-structured lines with proper indentation
-/// `parent_has_more` indicates if there are more items after parameters (response, exit_code)
-fn format_tool_parameters(
-    params: &serde_json::Value,
-    prefix: &str,
-    parent_has_more: bool,
-    content_width: usize,
-) -> Vec<Line<'static>> {
-    let mut lines = Vec::new();
-
-    match params {
-        serde_json::Value::Object(map) if map.is_empty() => {
-            lines.push(Line::from(Span::styled(
-                format!("{} parameters: (none)", prefix),
-                Style::default().fg(Color::DarkGray),
-            )));
-        }
-        serde_json::Value::Object(map) => {
-            // First line: parameters label
-            lines.push(Line::from(Span::styled(
-                format!("{} parameters:", prefix),
-                Style::default().fg(Color::DarkGray),
-            )));
-
-            // Parent vertical bar if there are more items after parameters
-            let parent_bar = if parent_has_more { "│" } else { " " };
-
-            // Each parameter on its own line with proper indentation
-            for (idx, (key, value)) in map.iter().enumerate() {
-                let is_last_param = idx == map.len() - 1;
-                let tree_char = if is_last_param { "└─" } else { "├─" };
-                let continuation_char = if is_last_param { "  " } else { "│ " };
-
-                // Use recursive formatter for all value types
-                let value_lines = format_json_value(
-                    value,
-                    parent_bar,
-                    tree_char,
-                    continuation_char,
-                    Some(key),
-                    content_width,
-                    0, // Start at depth 0
-                );
-                lines.extend(value_lines);
-            }
-        }
-        _ => {
-            lines.push(Line::from(Span::styled(
-                format!("{} parameters: {}", prefix, params),
-                Style::default().fg(Color::DarkGray),
-            )));
-        }
-    }
-
-    lines
-}
-
-/// Format response content as JSON key-values (just like parameters)
-/// Returns tree-structured lines with proper indentation
-fn format_response_content(
-    content: &str,
-    prefix: &str,
-    parent_has_more: bool,
-    content_width: usize,
-) -> Vec<Line<'static>> {
-    let mut lines = Vec::new();
-
-    if content.is_empty() {
-        lines.push(Line::from(Span::styled(
-            format!("{} response: (empty)", prefix),
-            Style::default().fg(Color::DarkGray),
-        )));
-        return lines;
-    }
-
-    // Try to parse as JSON
-    if let Ok(json_value) = serde_json::from_str::<serde_json::Value>(content) {
-        // Response is JSON - format like parameters
-        lines.push(Line::from(Span::styled(
-            format!("{} response:", prefix),
-            Style::default().fg(Color::DarkGray),
-        )));
-
-        // Parent vertical bar if there are more items after response
-        let parent_bar = if parent_has_more { "│" } else { " " };
-
-        // Format each JSON field using recursive formatter
-        if let serde_json::Value::Object(map) = json_value {
-            for (idx, (key, value)) in map.iter().enumerate() {
-                let is_last_field = idx == map.len() - 1;
-                let tree_char = if is_last_field { "└─" } else { "├─" };
-                let continuation_char = if is_last_field { "  " } else { "│ " };
-
-                // Use recursive formatter for all value types (handles nested objects/arrays)
-                let value_lines = format_json_value(
-                    value,
-                    parent_bar,
-                    tree_char,
-                    continuation_char,
-                    Some(key),
-                    content_width,
-                    0, // Start at depth 0
-                );
-                lines.extend(value_lines);
-            }
-        }
-    } else {
-        // Plain text response - show as-is with wrapping
-        let parent_bar = if parent_has_more { "│" } else { " " };
-        let first_line_prefix = format!("{} ", prefix);
-        let continuation_prefix = format!("{}   ", parent_bar);
-        let wrapped_lines = wrap_value_with_indent(
-            content,
-            &first_line_prefix,
-            &continuation_prefix,
-            content_width,
-        );
-
-        for wrapped_line in wrapped_lines {
-            lines.push(Line::from(Span::styled(
-                wrapped_line,
-                Style::default().fg(Color::DarkGray),
-            )));
-        }
-    }
-
-    lines
-}
-
-/// Truncate output for display (keep last N chars)
-#[allow(dead_code)] // Utility for tool output display
-fn truncate_output(output: &str, max_chars: usize) -> String {
-    if output.len() <= max_chars {
-        output.to_string()
-    } else {
-        format!("...{}", &output[output.len() - max_chars..])
-    }
-}
-
-/// Normalize line endings to Unix style (\n only)
-/// Converts Windows (\r\n) and old Mac (\r) line endings to Unix (\n)
-fn normalize_line_endings(text: &str) -> String {
-    // Replace \r\n with \n first (Windows style)
-    // Then replace any remaining \r with \n (old Mac style)
-    text.replace("\r\n", "\n").replace('\r', "\n")
-}
-
-/// Wrap text content with proper indentation for continuation lines
-/// Returns a vector of strings, each representing a wrapped line
-fn wrap_with_indent(
-    content: &str,
-    icon_width: usize,
-    content_width: usize,
-    subsequent_indent: &str,
-) -> Vec<String> {
-    use textwrap::{wrap, Options};
-
-    if content.is_empty() {
-        return vec![];
-    }
-
-    // Content width available (accounting for icon on first line)
-    let effective_width = content_width.saturating_sub(icon_width).max(10);
-
-    let options = Options::new(effective_width)
-        .initial_indent("")
-        .subsequent_indent(subsequent_indent)
-        .break_words(true);
-
-    wrap(content, options)
-        .into_iter()
-        .map(|cow| cow.into_owned())
-        .collect()
-}
-
-/// Calculate visual height of lines accounting for wrapping
-/// This matches how ratatui's Paragraph widget wraps text
-fn calculate_wrapped_height(lines: &[Line], content_width: usize) -> usize {
-    let mut height = 0;
-    for line in lines {
-        let line_width: usize = line.spans.iter().map(|span| span.content.width()).sum();
-        let wrapped_lines = if line_width == 0 {
-            1
-        } else {
-            line_width.div_ceil(content_width)
-        };
-        height += wrapped_lines;
-    }
-    height
-}
-
-/// Format user message (icon column + continuation indent)
-fn format_user_message(message: &Message, content_width: usize) -> Vec<Line<'static>> {
-    const ICON: &str = "$ ";
-    const INDENT: &str = "  ";
-
-    let mut result = Vec::new();
-
-    if message.content.is_empty() {
-        // Empty message - just show icon
-        result.push(Line::from(Span::styled(
-            ICON,
-            Style::default()
-                .fg(Color::Cyan)
-                .add_modifier(Modifier::BOLD),
-        )));
-        result.push(Line::from(""));
-        return result;
-    }
-
-    // Normalize line endings first to handle \r\n and bare \r
-    let normalized_content = normalize_line_endings(&message.content);
-
-    // Handle each paragraph (newline-separated) separately
-    let mut is_first_line_overall = true;
-    let paragraphs: Vec<&str> = normalized_content.lines().collect();
-
-    for paragraph in paragraphs.iter() {
-        // Empty paragraphs represent blank lines (from \n\n in content)
-        // Push an empty Line to create visual blank line
-        if paragraph.is_empty() {
-            result.push(Line::from(""));
-            continue;
-        }
-
-        // Wrap this paragraph with proper indentation
-        let wrapped = wrap_with_indent(paragraph, ICON.len(), content_width, INDENT);
-
-        for (line_idx, line_content) in wrapped.iter().enumerate() {
-            if is_first_line_overall && line_idx == 0 {
-                // Very first line: icon + content
-                result.push(Line::from(vec![
-                    Span::styled(
-                        ICON,
-                        Style::default()
-                            .fg(Color::Cyan)
-                            .add_modifier(Modifier::BOLD),
-                    ),
-                    Span::raw(line_content.clone()),
-                ]));
-                is_first_line_overall = false;
-            } else if line_idx == 0 {
-                // First line of non-first paragraph: needs indent
-                result.push(Line::from(format!("{}{}", INDENT, line_content)));
-            } else {
-                // Continuation: already indented by wrap_with_indent
-                result.push(Line::from(line_content.clone()));
-            }
-        }
-    }
-
-    // Add blank line separator between messages
-    result.push(Line::from(""));
-    result
-}
-
-/// Format LLM/assistant message (icon column + continuation indent)
-fn format_assistant_message(
-    message: &Message,
-    throbber: char,
-    content_width: usize,
-) -> Vec<Line<'static>> {
-    const INDENT: &str = "  ";
-
-    let (icon, icon_style) = if message.streaming {
-        (
-            format!("{} ", throbber),
-            Style::default()
-                .fg(RUST_ORANGE)
-                .add_modifier(Modifier::BOLD),
-        )
-    } else {
-        (
-            "● ".to_string(),
-            Style::default()
-                .fg(Color::Green)
-                .add_modifier(Modifier::BOLD),
-        )
-    };
-
-    let mut result = Vec::new();
-
-    if message.content.is_empty() {
-        // Empty message - just show icon
-        result.push(Line::from(Span::styled(icon, icon_style)));
-        result.push(Line::from(""));
-        return result;
-    }
-
-    // Normalize line endings first to handle \r\n and bare \r
-    let normalized_content = normalize_line_endings(&message.content);
-
-    // Handle each paragraph (newline-separated) separately
-    let mut is_first_line_overall = true;
-    let paragraphs: Vec<&str> = normalized_content.lines().collect();
-
-    for paragraph in paragraphs.iter() {
-        // Empty paragraphs represent blank lines (from \n\n in content)
-        // Push an empty Line to create visual blank line
-        if paragraph.is_empty() {
-            result.push(Line::from(""));
-            continue;
-        }
-
-        // Wrap this paragraph with proper indentation
-        let wrapped = wrap_with_indent(paragraph, icon.len(), content_width, INDENT);
-
-        for (line_idx, line_content) in wrapped.iter().enumerate() {
-            if is_first_line_overall && line_idx == 0 {
-                // Very first line: icon + content
-                result.push(Line::from(vec![
-                    Span::styled(icon.clone(), icon_style),
-                    Span::raw(line_content.clone()),
-                ]));
-                is_first_line_overall = false;
-            } else if line_idx == 0 {
-                // First line of non-first paragraph: needs indent
-                result.push(Line::from(format!("{}{}", INDENT, line_content)));
-            } else {
-                // Continuation: already indented by wrap_with_indent
-                result.push(Line::from(line_content.clone()));
-            }
-        }
-    }
-
-    // Add blank line separator between messages
-    result.push(Line::from(""));
-    result
-}
-
-/// Format system message (icon column + continuation indent)
-fn format_system_message(message: &Message, content_width: usize) -> Vec<Line<'static>> {
-    const INDENT: &str = "  ";
-
-    let icon = if message.collapsible {
-        if message.collapsed {
-            "▶ "
-        } else {
-            "▼ "
-        }
-    } else {
-        "→ "
-    };
-
-    let content = message.display_content();
-    let style = Style::default()
-        .fg(Color::Gray)
-        .add_modifier(Modifier::ITALIC);
-
-    let mut result = Vec::new();
-
-    if content.is_empty() {
-        // Empty message - just show icon
-        result.push(Line::from(Span::styled(
-            icon,
-            Style::default().fg(Color::Gray),
-        )));
-        result.push(Line::from(""));
-        return result;
-    }
-
-    // Normalize line endings first to handle \r\n and bare \r
-    let normalized_content = normalize_line_endings(content);
-
-    // Handle each paragraph (newline-separated) separately
-    let mut is_first_line_overall = true;
-    let paragraphs: Vec<&str> = normalized_content.lines().collect();
-
-    for paragraph in paragraphs.iter() {
-        // Empty paragraphs represent blank lines (from \n\n in content)
-        // Push an empty Line to create visual blank line
-        if paragraph.is_empty() {
-            result.push(Line::from(""));
-            continue;
-        }
-
-        // Wrap this paragraph with proper indentation
-        let wrapped = wrap_with_indent(paragraph, icon.len(), content_width, INDENT);
-
-        for (line_idx, line_content) in wrapped.iter().enumerate() {
-            if is_first_line_overall && line_idx == 0 {
-                // Very first line: icon + content
-                result.push(Line::from(vec![
-                    Span::styled(icon, Style::default().fg(Color::Gray)),
-                    Span::styled(line_content.clone(), style),
-                ]));
-                is_first_line_overall = false;
-            } else if line_idx == 0 {
-                // First line of non-first paragraph: needs indent
-                result.push(Line::from(Span::styled(
-                    format!("{}{}", INDENT, line_content),
-                    style,
-                )));
-            } else {
-                // Continuation: already indented by wrap_with_indent
-                result.push(Line::from(Span::styled(line_content.clone(), style)));
-            }
-        }
-    }
-
-    // Add blank line separator between messages
-    result.push(Line::from(""));
-    result
 }
 
 fn render_messages(frame: &mut Frame, area: Rect, app: &mut App, throbber: char) -> usize {
@@ -1127,12 +506,8 @@ fn render_messages(frame: &mut Frame, area: Rect, app: &mut App, throbber: char)
             }
 
             // Normal message rendering (non-tool messages)
-            // Use new formatting functions based on message role
-            let message_lines = match message.role {
-                Role::User => format_user_message(message, content_width),
-                Role::Assistant => format_assistant_message(message, throbber, content_width),
-                Role::System => format_system_message(message, content_width),
-            };
+            // Use formatting functions from message_formatter module
+            let message_lines = format_message_by_role(message, throbber, content_width);
 
             // Add formatted lines to buffer
             // Trailing blank separator will be removed globally after loop
